+++ src/datasets/resource_test_scripts/pancreas_tasks.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-src/denoising/resources_test_scripts/pancreas.sh
-src/dimensionality_reduction/resources_test_scripts/pancreas.sh
-src/label_projection/resources_test_scripts/pancreas.sh
+src/tasks/denoising/resources_test_scripts/pancreas.sh
+src/tasks/dimensionality_reduction/resources_test_scripts/pancreas.sh
+src/tasks/label_projection/resources_test_scripts/pancreas.sh