+++ src/tasks/batch_integration/resources_test_scripts/pancreas.sh
@@ -18,23 +18,23 @@ mkdir -p $DATASET_DIR
 
 # process dataset
 echo process data...
-viash run src/batch_integration/process_dataset/config.vsh.yaml -- \
+viash run src/tasks/batch_integration/process_dataset/config.vsh.yaml -- \
   --input $RAW_DATA \
   --output $DATASET_DIR/unintegrated.h5ad \
   --hvgs 100
 
 echo Running BBKNN
-viash run src/batch_integration/methods/bbknn/config.vsh.yaml -- \
+viash run src/tasks/batch_integration/methods/bbknn/config.vsh.yaml -- \
   --input $DATASET_DIR/unintegrated.h5ad \
   --output $DATASET_DIR/bbknn.h5ad
 
 echo Running SCVI
-viash run src/batch_integration/methods/scvi/config.vsh.yaml -- \
+viash run src/tasks/batch_integration/methods/scvi/config.vsh.yaml -- \
   --input $DATASET_DIR/unintegrated.h5ad \
   --output $DATASET_DIR/scvi.h5ad
 
 echo Running combat
-viash run src/batch_integration/methods/combat/config.vsh.yaml -- \
+viash run src/tasks/batch_integration/methods/combat/config.vsh.yaml -- \
   --input $DATASET_DIR/unintegrated.h5ad \
   --output $DATASET_DIR/combat.h5ad