+++ src/tasks/denoising/resources_scripts/run_benchmark.sh
@@ -64,7 +64,7 @@ fi
 export NXF_VER=22.04.5
 nextflow \
   run . \
-  -main-script src/denoising/workflows/run/main.nf \
+  -main-script src/tasks/denoising/workflows/run/main.nf \
   -profile docker \
   -params-file "$params_file" \
   --publish_dir "$OUTPUT_DIR" \