+++ src/tasks/label_projection/workflows/run/run_test.sh
@@ -22,7 +22,7 @@ export NXF_VER=22.04.5
 
 nextflow \
   run . \
-  -main-script src/label_projection/workflows/run/main.nf \
+  -main-script src/tasks/label_projection/workflows/run/main.nf \
   -profile docker \
   -resume \
   --id pancreas \